bugfix: sophie_links_siblings_of not working
Prior to this fix, the tag only worked for sophie_entry_list_url page
due to its assumption that the format of the reversed url is fixed.

But generally the url contains a slug certain part of the blog,
as apposed to sophie_entry_list_url, which contained not such.