added complimentary methods to URLStructure.
When URLStructure.pagination_suffix is set to be true,
urls defined in urlpatterns will be suffixed with page_slug field.
Similarly, when URLStructure.blog_prefix is set to be true,
urls in urlpatterns will be prefixed with the blog_slug field if
necessary.

URLStructure.pagination_suffix is False by default while
URLStructure.blog_prefix is True by default.